+++ llm_assistant/tests/__init__.py
@@ -1,2 +1,3 @@
 from . import test_prompt_arguments
 from . import test_loop_control
+from . import test_pause